Add LLRP support
Uses sllurp to implement a fake LLRP server.

This code currently works well enough that it can get past the
WebScorer connection test, and even puts some tags in the race
screen.

However, it's all hard-coded. The next step is to hook up to our
reader using HRP and feed in live data.