Add LCD support
Adds support to display messages on an I2C LCD screen,
for when LLRP2HRP runs on an embedded system (eg Raspberry Pi).

The screen will attempt to determine a usable IP address for connecting
to the reader and display it on the first line of an LCD.
Status messages will appear on the second line.